git-svn-id: svn://192.168.2.130:8807/sd/java/financialSystem@60859 ced6f03c-c1bb-49ab-ae0c-4da96410de7b

+++ src/main/java/cn/financial/service/impl/ResourceServiceImpl.java
@@ -1,13 +1,19 @@
 package cn.financial.service.impl;
 
+import java.util.ArrayList;
 import java.util.List;
 
+import org.apache.commons.collections.CollectionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.alibaba.fastjson.JSONObject;
+
 import cn.financial.dao.ResourceDAO;
 import cn.financial.model.Resource;
+import cn.financial.model.RoleResource;
 import cn.financial.service.ResourceService;
+import cn.financial.util.TreeNode;
 
 
 @Service("ResourceServiceImpl")
@@ -33,6 +39,12 @@ public class ResourceServiceImpl implements ResourceService{
      */
     @Override
     public Integer insertResource(Resource resource) {
+        if(resource.getName() == null || resource.getName().equals("")){
+            return -1;
+        }
+        if(resource.getPermssion() == null || resource.getPermssion().equals("")){
+            return -2;
+        }
         return resourceDao.insertResource(resource);
     }
     /**
@@ -47,8 +59,47 @@ public class ResourceServiceImpl implements ResourceService{
      */
     @Override
     public Integer deleteResource(String resourceId) {
+        if(resourceId == null || resourceId.equals("")){
+            return -1;
+        }
         return resourceDao.deleteResource(resourceId);
     }
+    
+    //角色权限关联构建成树(勾选子节点不用勾选父节点)
+    /*public JSONObject resourceList(List<RoleResource> roleResource){
+        List<TreeNode<Resource>> nodes = new ArrayList<>();
+        List<Resource> listre = new ArrayList<>();
+        for (int i = 0; i < roleResource.size(); i++) {
+            String resourceId =roleResource.get(i).getsId();
+            //当前的resource节点
+            Resource resource = resourceDao.getResourceById(resourceId, null);
+            listre.add(resource);
+            String[] split = resource.getParentId().split("/");
+            for (int j = 0; j < split.length; j++) {
+                Resource resource1 = resourceDao.getResourceById(null, split[j]);
+                listre.add(resource1);
+            }
+        }
+        if(!CollectionUtils.isEmpty(listre)){
+            for (int i = 0; i < listre.size() - 1; i++) {
+                for (int j = listre.size() - 1; j > i; j--) {
+                    if (listre.get(j).getId().equals(listre.get(i).getId())) {
+                        listre.remove(j);
+                    }
+                }
+            }  
+            for (Resource resource : listre) {
+                TreeNode<Resource> node = new TreeNode<>();
+                node.setId(resource.getCode().toString());
+                String b=resource.getParentId().substring(resource.getParentId().lastIndexOf("/")+1);
+                node.setParentId(b);
+                node.setName(resource.getName());
+                nodes.add(node);
+            }
+        }
+        JSONObject jsonObject = (JSONObject) JSONObject.toJSON(TreeNode.buildTree(nodes));
+        return jsonObject;
+    }*/
     
 }
